+++ core/jobs/src/test/java/com/cognifide/aet/job/common/modifiers/executejavascript/ExecuteJavaScriptModifierTest.java
@@ -29,6 +29,8 @@ import org.openqa.selenium.WebDriver;
 
 import java.util.Map;
 
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.Mockito.when;
 
 
@@ -36,9 +38,8 @@ import static org.mockito.Mockito.when;
 public class ExecuteJavaScriptModifierTest {
 
     private static final String PARAM_CMD = "cmd";
-
+    private static final String PARAM_URL = "snippetUrl";
     private static final String PARAM_CMD_VALUE = "document.getElementById('toRemove').style.display='none'";
-
     private static final String URL = "http://www.cognifide.com";
 
     @Mock
@@ -56,17 +57,67 @@ public class ExecuteJavaScriptModifierTest {
     }
 
     @Test
-    public void setParameters_CmdIsValid_ValidationPassedSuccessfuly() throws ParametersException {
-        when(params.containsKey(PARAM_CMD)).thenReturn(true);
-        when(params.get(PARAM_CMD)).thenReturn(PARAM_CMD_VALUE);
+    public void encodeBasicAuth_ComplexPassword_ShouldBeProperlyEncoded() {
+        String actual = tested.encodeBasicAuth("admin", "zażółć!@#$%^&*()_+-={}|[]\\:\";'<>?,./");
+        String base64Encoded = "YWRtaW46emHFvMOzxYLEhyFAIyQlXiYqKClfKy09e318W11cOiI7Jzw+PywuLw==";
+        assertThat(actual, equalTo(base64Encoded));
+    }
+
+    @Test
+    public void encodeBasicAuth_EmptyPassword_ShouldBeProperlyEncoded() {
+        String actual = tested.encodeBasicAuth("admin", "");
+        String base64Encoded = "YWRtaW46";
+        assertThat(actual, equalTo(base64Encoded));
+    }
+
+    @Test
+    public void encodeBasicAuth_NullPassword_ShouldBeTreatedAsEmptyString() {
+        String actual = tested.encodeBasicAuth("admin", null);
+        String base64Encoded = "YWRtaW46";
+        assertThat(actual, equalTo(base64Encoded));
+    }
+
+    @Test(expected = ParametersException.class)
+    public void setParameters_MissingParam_ValidationPassedUnsuccessfully() throws ParametersException {
+        when(params.get("basicAuthUsername")).thenReturn("admin");
+        when(params.get("basicAuthPassword")).thenReturn(null);
+
+        tested.setParameters(params);
+    }
+
+    @Test
+    public void setParameters_CmdIsValid_ValidationPassedSuccessfully() throws ParametersException {
+        setupCmdParam();
+        tested.setParameters(params);
+    }
+
+    @Test
+    public void setParameters_SnippetUrlIsValid_ValidationPassedSuccessfully() throws ParametersException {
+        setupSnippetUrlParam();
+        tested.setParameters(params);
+    }
+
+    @Test
+    public void setParameters_BothCmdAndSnippetUrlAreProvided_ValidationPassedSuccessfully() throws ParametersException {
+        setupCmdParam();
+        setupSnippetUrlParam();
         tested.setParameters(params);
     }
 
     @Test(expected = ParametersException.class)
-    public void setParameters_CmdIsNotPassed_ValidationPassedUnsuccessfuly()
-            throws ParametersException {
+    public void setParameters_NoSourceIsPassed_ValidationPassedUnsuccessfully()
+        throws ParametersException {
         tested.setParameters(params);
     }
 
+    private void setupSnippetUrlParam() {
+        when(params.containsKey(PARAM_URL)).thenReturn(true);
+        when(params.get(PARAM_URL)).thenReturn(URL);
+    }
+
+    private void setupCmdParam() {
+        when(params.containsKey(PARAM_CMD)).thenReturn(true);
+        when(params.get(PARAM_CMD)).thenReturn(PARAM_CMD_VALUE);
+    }
 
 }